Repaired Visual Studio solution for x86
- added x86 configuration
- added x86 and x64 sdl3.dlls to solution
- fixed x86 compilation warnings

+++ sdl_main/sdl_main.cpp
@@ -271,7 +271,7 @@ SDL_AppResult SDL_AppEvent(void *appstate, SDL_Event *event)
         {
             int nextVideoFilterIndex = currentVideoFilterIndex + 1;
 
-            if (nextVideoFilterIndex >= videoFilters.size())
+            if (nextVideoFilterIndex >= (int)videoFilters.size())
                 nextVideoFilterIndex = 0;
 
             selectFilter(nextVideoFilterIndex);
